Fixed issue previously unseen on main computor
Fixed an issue that didn't turn up on my main computor, but was clear in the results when running the program on my laptop where using mod instead of fract when generating noise caused issues with very specific values of scale. Namely, when using scale = 50.0f there were huge seems that showed up out of nowhere, but were not present on my desktop.