Refactored so that App is no longer a god class
Made it so the app simply runs the gameloop and broke its systems into a ChunkManager and a Renderer.

I still need to break up the Renderer class since it is doing a bit to much. My current goal is to break it into three distinct classes:

1. Shader Manager - handles shader setup logic and stores shaders.

2. Window Manager - Handles the creation of the glfw/glad window.

3. ImGui Manager - Handles all the ImGui logic.

Afterwards I'll expand the library to include marching cubes algorithm created on the GPU and I'll create a demo for that as well.