Moved things around and made the library save the programs.
Made the library save one version of each program as opposed to before where it was remade for every chunk that tried to render. Also have moved things around to try and accomodate multithreading, but since the compute shaders are using openGL they need a glfw context, which they can't have at the same time as the render thread, meaning that multithreading is only really applicable for using the cpu and not for using opengl computeShaders in combination with the opengl vertex and fragment shaders.