Moved the FlatTriTable away from GPU
After user testing it was found that their graphics card was not able to use a 4096 large array as a global glsl variable. Therefore I am now sending the data from CPU as a buffer to the GPU to be used, which was the original plan but is a bit more heavy on the machine.