Fixed a few bugs and quirks
Some bugs that I got help identefying during playtesting were squashed, but some of them remain in other demos which I will fix tomorrow. Also, I am aware of a bug where memory is not dealocated when remeshing the terrain. This might have more of a drawback for the demo than any practical issues when using the library, since removing every saved chunk is not very likely to be of any use normally, but is a nice feature for the demo and should not be something that stays in the final product anyway.