Rethough my approach to noise with chunks
I am no longer using perlin noise with an FBM approach since it makes a few things difficult or impossible. First, I cant really use a particularly great approach to splines if I use FBM with perlin noise since the range of the noise is greater and the peaks are way further away from the avarages making control harder. What I did to try and alleviate this issue was to use contrast stretching ((f(x,y,z)-minValue)/(maxValue-minValue)) but this doesnt function in the project as it will make neighbouring chunks incoherrant. Therefore, it makes more sense to simply not use FBM and use noise that is defined within a range. What I did was to copy an implementation of simplex noise which yields VERY similar results as FBM perlin noise. The noise is almost always defined evenly within [-0.9,0.9], which allows us to contrast stretch without additional passes or have it be incongruent with neighbouring chunks. When this issue was solved, I added in caves since it became much easier.