Added a padded texture to reduce aliasing and terrain paint addiotions
Added padding to each block texture in the pallete as to not have bleeding and aliasing issues. There still are issues, but they are far fewer. Also added some more logic to terrain painting to add in dirt then stone at parts further down under the ground. This is not very procedural and will be updated at some point.